Turned "not contiguous or the first atom" warning into error.
Issuing warning in the program is hardly respected (even though current warn() changes color).
Previously I left the error message as warning because it is useful for experts.
But experts can edit the source code by themselves or read comments, thus no reason to keep this as a warning now.